+++ apps/editor-dev/pages/plugins/ToolbarPlugin.jsx
@@ -1,7 +1,8 @@
 import DarkModeFilled from "@elf-framework/icon/DarkModeFilled";
 import LightModeFilled from "@elf-framework/icon/LightModeFilled";
 import PolylineFilled from "@elf-framework/icon/PolylineFilled";
-import { Toolbar } from "@elf-framework/ui";
+import { useComponentRender } from "@elf-framework/sapa";
+import { ActionGroup, Button, Toolbar } from "@elf-framework/ui";
 
 export function ToolbarPlugin(editor) {
   const leftMenu = {
@@ -69,12 +70,54 @@ export function ToolbarPlugin(editor) {
     ],
   };
 
+  function CustomMenuItem() {
+    useComponentRender("config:editor.layout.show.left");
+    useComponentRender("config:editor.layout.show.right");
+
+    const isShowLeft = editor.configs.get("editor.layout.show.left");
+    const isShowRight = editor.configs.get("editor.layout.show.right");
+
+    return (
+      <ActionGroup compact>
+        <Button
+          size="small"
+          variant={isShowLeft ? "primary" : "default"}
+          iconOnly
+          selected={isShowLeft}
+          onClick={() => {
+            editor.commands.execute("editor.layout.toggle.left");
+          }}
+        >
+          L
+        </Button>
+        <Button
+          size="small"
+          variant={isShowRight ? "primary" : "default"}
+          iconOnly
+          selected={isShowRight}
+          onClick={() => {
+            editor.commands.execute("editor.layout.toggle.right");
+          }}
+        >
+          R
+        </Button>
+      </ActionGroup>
+    );
+  }
+
   const centerMenu = {
     items: [
       {
         type: "item",
         title: "hoverable false tools item",
         hoverable: false,
+      },
+      {
+        type: "custom",
+        hoverable: false,
+        render: () => {
+          return <CustomMenuItem />;
+        },
       },
     ],
   };