+++ packages/base-editor/src/commands/editor.layout.toggle.right.js
@@ -1,0 +1,7 @@
+export default {
+  command: "editor.layout.toggle.right",
+  execute: function (editor) {
+    var isShow = editor.configs.get("editor.layout.show.right");
+    editor.configs.set("editor.layout.show.right", !isShow);
+  },
+};