+++ packages/base-editor/src/commands/editor.layout.toggle.top.js
@@ -1,0 +1,7 @@
+export default {
+  command: "editor.layout.toggle.top",
+  execute: function (editor) {
+    var isShow = editor.configs.get("editor.layout.show.top");
+    editor.configs.set("editor.layout.show.top", !isShow);
+  },
+};