+++ packages/base-editor/src/configs/editor.layout.show.bottom.js
@@ -1,0 +1,7 @@
+export default {
+  key: "editor.layout.show.bottom",
+  defaultValue: true,
+  title: "Show bottom panel in layout ",
+  description: "Show bottom panel in layout",
+  type: "boolean",
+};