+++ packages/base-editor/src/configs/editor.layout.show.left.js
@@ -1,0 +1,7 @@
+export default {
+  key: "editor.layout.show.left",
+  defaultValue: true,
+  title: "Show left panel in layout ",
+  description: "Show left panel in layout",
+  type: "boolean",
+};