+++ packages/base-editor/src/configs/editor.layout.show.right.js
@@ -1,0 +1,7 @@
+export default {
+  key: "editor.layout.show.right",
+  defaultValue: true,
+  title: "Show right panel in layout ",
+  description: "Show right panel in layout",
+  type: "boolean",
+};