+++ packages/base-editor/src/configs/editor.layout.show.top.js
@@ -1,0 +1,7 @@
+export default {
+  key: "editor.layout.show.top",
+  defaultValue: true,
+  title: "Show top panel in layout ",
+  description: "Show top panel in layout",
+  type: "boolean",
+};