+++ packages/base-editor/src/configs/index.js
@@ -1,3 +1,16 @@
+import editorLayoutShowBottom from "./editor.layout.show.bottom";
+import editorLayoutShowLeft from "./editor.layout.show.left";
+import editorLayoutShowRight from "./editor.layout.show.right";
+import editorLayoutShowTop from "./editor.layout.show.top";
 import editorTheme from "./editor.theme";
 
-export default [editorTheme];
+export default [
+  // theme
+  editorTheme,
+
+  // layout
+  editorLayoutShowTop,
+  editorLayoutShowLeft,
+  editorLayoutShowRight,
+  editorLayoutShowBottom,
+];