+++ packages/sapa/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @elf-framework/sapa
 
+## 0.0.107
+
+### Patch Changes
+
+- update index.d.ts
+
 ## 0.0.85
 
 ### Patch Changes